+++ .gitignore
@@ -6,3 +6,4 @@ __pycache__
 *.mp3
 output
 .aider*
+.env
